\name{PCB}
\alias{PCB}
\docType{data}
\title{concentration of polychlorinated biphenyl (PCB) in the yolk lipids of pelican eggs}
\description{Concentration of polychlorinated biphenyl (PCB) in the yolk lipids of pelican eggs, published by Risebrough (1972) and 
used by Thas (2010).
}
\usage{data(PCB)}
\format{
The data are in one vector, named PCB.
}
\source{Risebrough (1972) \emph{Effects of environmental pollutants upon animals other than man}, Proceedings of the 6th Berkeley Symposium on Mathematical Statistics and Probability 443--463 
}
\references{Used by Thas (2010) \emph{Comparing Distributions}, Springer. 978-0-387-92710-7
}
\examples{
data(PCB)
}
\keyword{datasets}
