\name{PCB}
\alias{PCB}
\alias{PCB1}
\docType{data}
\title{concentration of polychlorinated biphenyl (PCB) in the yolk lipids of pelican eggs in ppm}
\description{Concentration (parts per million) of 
polychlorinated biphenyl (PCB) in the yolk lipids of pelican 
(\emph{Pelecanus occidentalis}) eggs, from Anacapa, California.
Published by Risebrough (1972).

The data \code{PCB} are the data as published in Risebrough (1972).

These data are also used in Thas (2010), as table 1, with 
a difference in the 24th observation (265 in Risebrough and
256 in Thas).  \code{PCB1} is the version in table 1 of Thas (2010).

This is a subset of a wider collection of data on \emph{Pelecanus 
occidentalis} eggs from Risebrough et al (undated).
}
\usage{data(PCB)}
\format{
The data are in one vector, named PCB.
}
\source{Risebrough (1972) \emph{Effects of environmental pollutants upon animals other than man}, Proceedings of the 6th Berkeley Symposium on Mathematical Statistics and Probability, Volume 6: Effects of Pollution on Health, 443--463, University of California Press, Berkeley, California.

http://projecteuclid.org/euclid.bsmsp/1200514718

R.W. Risebrough, F. Gress, J.K. Baptista, D.W. Anderson and
R.W Schreiber (undated) \emph{Oceanic pollution: Effects on the
	reproduction of Brown Pelicans Pelecanus occidentalis}, unpublished manuscript
}
\references{Thas (2010) \emph{Comparing Distributions}, Springer. 978-0-387-92710-7
}
\examples{
data(PCB)
}
\keyword{datasets}
